Moved tests used in Range widgets to their base classes
